requirements.txt
Django==1.10.3
django-sesame==1.2
